pin langugae to c23

+++ TOMB4/game/actorme.h
@@ -1,7 +1,7 @@
 #ifndef ACTORME_INCLUDED
 #define ACTORME_INCLUDED
 typedef struct ACTORME {
-	long offset;
+	int offset;
 	short objslot;
 	short nodes;
 } ACTORME;
