+++ TOMB4/game/aiinfo.h
@@ -3,9 +3,9 @@
 typedef struct AI_INFO {
 	short zone_number;
 	short enemy_zone;
-	long distance;
-	long ahead;
-	long bite;
+	int distance;
+	int ahead;
+	int bite;
 	short angle;
 	short x_angle;
 	short enemy_facing;
