+++ TOMB4/game/aiobject.h
@@ -4,9 +4,9 @@
 typedef struct AIOBJECT {
 	short object_number;
 	short room_number;
-	long x;
-	long y;
-	long z;
+	int x;
+	int y;
+	int z;
 	short trigger_flags;
 	short flags;
 	short y_rot;
