+++ TOMB4/game/bikeinfo.h
@@ -3,15 +3,15 @@
 typedef struct BIKEINFO {
 	short right_front_wheelrot;
 	short right_back_wheelrot;
-	long left_wheelrot;
-	long velocity;
-	long unused1;
-	long pitch1;
+	int left_wheelrot;
+	int velocity;
+	int unused1;
+	int pitch1;
 	short move_angle;
 	short extra_rotation;
 	short rot_thing;
-	long bike_turn;
-	long pitch2;
+	int bike_turn;
+	int pitch2;
 	short flags;
 	short light_intensity;
 } BIKEINFO;
