+++ TOMB4/game/biteinfo.h
@@ -1,9 +1,9 @@
 #ifndef TOMB4_GAME_BITEINFO_H
 #define TOMB4_GAME_BITEINFO_H
 typedef struct BITE_INFO {
-	long x;
-	long y;
-	long z;
-	long mesh_num;
+	int x;
+	int y;
+	int z;
+	int mesh_num;
 } BITE_INFO;
 #endif// TOMB4_GAME_BITEINFO_H
