+++ TOMB4/game/camerainfo.h
@@ -11,13 +11,13 @@ typedef struct CAMERA_INFO {
 	GAME_VECTOR target;
 	camera_type type;
 	camera_type old_type;
-	long shift;
-	long flags;
-	long fixed_camera;
-	long number_frames;
-	long bounce;
-	long underwater;
-	long target_distance;
+	int shift;
+	int flags;
+	int fixed_camera;
+	int number_frames;
+	int bounce;
+	int underwater;
+	int target_distance;
 	short target_angle;
 	short target_elevation;
 	short actual_elevation;
@@ -29,7 +29,7 @@ typedef struct CAMERA_INFO {
 	short speed;
 	ITEM_INFO* item;
 	ITEM_INFO* last_item;
-	long mike_at_lara;
+	int mike_at_lara;
 	PHD_VECTOR mike_pos;
 } CAMERA_INFO;
 #endif// TOMB4_GAME_CAMERAINFO_H
