+++ TOMB4/game/clockworkbeetle.h
@@ -1,7 +1,7 @@
 #ifndef TOMB4_GAME_CLOCKWORKBEETLE_H
 #define TOMB4_GAME_CLOCKWORKBEETLE_H
 typedef struct ITEM_INFO ITEM_INFO;
-ITEM_INFO* TriggerClockworkBeetle(long flag);
+ITEM_INFO* TriggerClockworkBeetle(int flag);
 void ControlClockworkBeetle(short item_number);
