+++ TOMB4/game/combinelist.h
@@ -1,7 +1,7 @@
 #ifndef TOMB4_GAME_COMBINELIST_H
 #define TOMB4_GAME_COMBINELIST_H
 typedef struct COMBINELIST {
-	void (*combine_routine)(long flag);
+	void (*combine_routine)(int flag);
 	short item1;
 	short item2;
 	short combined_item;
